fix: workaround stupid compiler
