+++ src/features/token-migration/components/migration-details/__tests__/migration-details.test.tsx
@@ -5,7 +5,8 @@ import { useParams } from "react-router-dom"
 import { useSingleTransactionList } from "features/transactions/queries"
 import { useGetBlock } from "features/network/queries"
 import { createMockTxList, mockSingleTxListError } from "test/transactions"
-import { createMockBlock, mockBlockError } from "test/block"
+import { mockBlockError } from "test/block"
+import { createMockBlock } from "features/token-migration/test-utils/mocks"
 
 const mockEventId = "6a9900000001"
 const mockTxHash = "012345"
