--- src/test/send.tsx
@@ -1,7 +1,0 @@
-import { useCreateSendTxn } from "../features/transactions"
-
-export const mockUseCreateSendTransaction = () => {
-  useCreateSendTxn.mockReturnValue({
-    mutateAsync: jest.fn().mockResolvedValue(true),
-  })
-}
