--- src/test/uuid.tsx
@@ -1,6 +1,0 @@
-import { randomUUID } from "crypto"
-
-export const mockRandomUUID = () => {
-  const mockUuid = "mockUUID"
-  randomUUID.mockReturnValue(mockUuid)
-}
